+++ lib/jss/server.rb
@@ -32,7 +32,7 @@ module JSS
   # attribute. It is created fresh every time {APIConnection#connect} is called.
   #
   # That's the only time it should be instantiated, and all access should be
-  # through {JSS::API.server}
+  # through {JSS.api_connection.server}
   #
   class Server
 
@@ -78,7 +78,7 @@ module JSS
       # However, it's marked as 'deprecated'. Hopefully jamf will
       # keep this basic level of info available for basic authentication
       # and JSS version checking.
-      ju = JSS::API.get_rsrc('jssuser')[:user]
+      ju = JSS.api_connection.get_rsrc('jssuser')[:user]
       @license_type = ju[:license_type]
       @product = ju[:product]
       @raw_version = ju[:version]
@@ -89,18 +89,18 @@ module JSS
       @version = parsed[:version]
 
     rescue RestClient::Request::Unauthorized
-      raise JSS::AuthenticationError, "Incorrect JSS username or password for '#{JSS::API.jss_user}@#{JSS::API.server_host}'."
+      raise JSS::AuthenticationError, "Incorrect JSS username or password for '#{JSS.api_connection.jss_user}@#{JSS.api_connection.server_host}'."
     end
 
     # @return [String] the organization to which the server is licensed
     def organization
-      @act_code_data ||= JSS::API.get_rsrc(ACTIVATION_CODE_RSRC)[ACTIVATION_CODE_KEY]
+      @act_code_data ||= JSS.api_connection.get_rsrc(ACTIVATION_CODE_RSRC)[ACTIVATION_CODE_KEY]
       @act_code_data[:organization_name]
     end
 
     # @return [String] the activation code for the server licence
     def activation_code
-      @act_code_data ||= JSS::API.get_rsrc(ACTIVATION_CODE_RSRC)[ACTIVATION_CODE_KEY]
+      @act_code_data ||= JSS.api_connection.get_rsrc(ACTIVATION_CODE_RSRC)[ACTIVATION_CODE_KEY]
       @act_code_data[:code]
     end